Make "positional" and "required" two orthogonal dimensions
In previous design, we have separate fields for marking types of a
field: positional and optional while only a single boolean will do.
There is no way of specifying an optional as required and we have
non-required positional which is not that consistent within itself

Instead of having the following and let users to memorize the order of
each argument

	<NAME> <LOADBALANCER> <{tcp,udp,http,https}> <LISTENER_PORT> <{rr,wrr,wlc,sch,tch}> <BACKEND_PORT>

Make it order-neutral with mnemonics

	 <--loadbalancer LOADBALANCER> <--listener-type {tcp,udp,http,https}> <--listener-port LISTENER_PORT> <--scheduler {rr,wrr,wlc,sch,tch}> <--backend-port BACKEND_PORT> <NAME>

This commit will bring the following changes

An argument is either positional or optional.  Optional in the sense
that the argument will be presented with `--` prefix, not the actual
argument must be required or non-required

An argument with field name being all UPPERCASE will be positional by
default, unless tagged with positional:false.

Positional argument must be "required", except in the case of the
MultiArgument where zero length is possible.  Optional argument will
have required:false by default and can be "required" explicitly